pyproject.toml
[project]
name = "fix-compile"
version = "0.2.0"
description = "Fix Docker build and runtime errors using LLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain[community,openai]>=1.2.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "httpx[socks]>=0.28.1",
]


[dependency-groups]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=4.5.1",
]


[project.scripts]
fix-compile = "fix_compile.main:main"

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"
